+++ src/sc_crawler/vendors/_azure.py
@@ -492,7 +492,9 @@ def _standardize_server(server: dict, vendor) -> dict:
     # }
     family, features, gpus, gpu_model, gpu_memory = _parse_server_name(server["name"])
     # override family from SKU listing
-    family = server["family"].removeprefix("standard").removesuffix("Family")
+    family = recompile(r"(?i)family$").sub(
+        "", recompile(r"(?i)^standard").sub("", server["family"])
+    )
 
     def capability(name: str, default=None) -> str:
         try:
@@ -552,13 +554,13 @@ def _standardize_server(server: dict, vendor) -> dict:
         "gpu_count": round(gpus, 4),
         "gpu_model": gpu_model,
         "gpu_memory_min": (
-            None
+            0
             if not (gpus and gpu_memory)
             else int(gpu_memory * gpus)
             if gpus <= 1
             else int(gpu_memory)
         ),
-        "gpu_memory_total": int(gpu_memory * gpus) if gpus and gpu_memory else None,
+        "gpu_memory_total": int(gpu_memory * gpus) if gpus and gpu_memory else 0,
         "storage_size": round(sum([s.size for s in storages])),  # int GB
         "storages": storages,
         # TODO: have to implement manual mapping for network_speed related fields